Add files via upload
This dashboard is created as part of the Forage's JP Morgan Chase & Co. Global Final and Business Management. The dataset consist of budget and time required for completing the projects with resources and respective tasks. The dataset is subjected to cleaning and exploration using Microsoft Power BI desktop. Page level filters, visual level filters, hierarchy and slicers were used to make the dashboard more appealing and interactive.